Move benchmark creation and time series to post-process
